started on team and tag architecture
this is a lot to get my head around
right  now I'm trying to get players to spawn in on a team and take on that team's colors without overwriting the other players colors... it's not going great. it looks like they are correctly saving their team info, so I think it's that they're all using the same material. I guess I need to make a new material for each team?? oof

also temporarily turned of the start of the Tag team swapping behavior while I get the above functioning. the start I have for this is running into a lot of weirdness with RPC calls being invalid, hopefully more reading and lecture will clear that up.